Simplify data cleaning script, save initial output
